Rename overlay visualizer types to avoid confusion
