--- Bonsai.Vision.Design/PolygonMashupVisualizer.cs
@@ -1,35 +1,0 @@
-﻿using Bonsai;
-using Bonsai.Design;
-using Bonsai.Vision.Design;
-using System;
-
-[assembly: TypeVisualizer(typeof(PolygonMashupVisualizer), Target = typeof(MashupSource<ImageMashupVisualizer, PolygonVisualizer>))]
-
-namespace Bonsai.Vision.Design
-{
-    /// <summary>
-    /// Provides a type visualizer that overlays a collection of polygonal regions
-    /// over an existing image visualizer.
-    /// </summary>
-    public class PolygonMashupVisualizer : DialogTypeVisualizer
-    {
-        ImageMashupVisualizer visualizer;
-
-        /// <inheritdoc/>
-        public override void Show(object value)
-        {
-            PolygonVisualizer.Draw(visualizer.VisualizerImage, value);
-        }
-
-        /// <inheritdoc/>
-        public override void Load(IServiceProvider provider)
-        {
-            visualizer = (ImageMashupVisualizer)provider.GetService(typeof(MashupVisualizer));
-        }
-
-        /// <inheritdoc/>
-        public override void Unload()
-        {
-        }
-    }
-}
