--- Bonsai.Vision.Design/RectMashupVisualizer.cs
@@ -1,35 +1,0 @@
-﻿using Bonsai;
-using Bonsai.Design;
-using Bonsai.Vision.Design;
-using System;
-
-[assembly: TypeVisualizer(typeof(RectMashupVisualizer), Target = typeof(MashupSource<ImageMashupVisualizer, RectVisualizer>))]
-
-namespace Bonsai.Vision.Design
-{
-    /// <summary>
-    /// Provides a type visualizer that overlays the visual representation of a
-    /// rectangle over an existing image visualizer.
-    /// </summary>
-    public class RectMashupVisualizer : DialogTypeVisualizer
-    {
-        ImageMashupVisualizer visualizer;
-
-        /// <inheritdoc/>
-        public override void Show(object value)
-        {
-            RectVisualizer.Draw(visualizer.VisualizerImage, value);
-        }
-
-        /// <inheritdoc/>
-        public override void Load(IServiceProvider provider)
-        {
-            visualizer = (ImageMashupVisualizer)provider.GetService(typeof(MashupVisualizer));
-        }
-
-        /// <inheritdoc/>
-        public override void Unload()
-        {
-        }
-    }
-}
